chore: only upload when hugo is successful

+++ toSk1Mserver.sh
@@ -1,4 +1,3 @@
 #!/bin/bash
 rm -rf public
-hugo
-rsync -acv --delete-delay public/ sk1m.de:/usr/share/nginx/de.edjopato/ | grep -E '^deleting|[^/]$|^$'
+hugo && rsync -acv --delete-delay public/ sk1m.de:/usr/share/nginx/de.edjopato/ | grep -E '^deleting|[^/]$|^$'
